requirements.txt
numpy
scipy
cryptography
requests
matplotlib
pyalsaaudio
